chore: Move parsing logic from module to definition (#181)
